Make it a policy to include the text of legal tools
In a future commit, I plan on adding a sound effect to this project. I’m
not going to make the sound effect myself. Instead, I’m going to find
one that’s either available under a free license or in the public
domain. It’s more likely that I’ll end up finding a freely licensed one
than a public domain one.

If I do end up finding something that’s available under a free license,
then that license will probably require that I include a copy of it. For
the sake of consistency, this commit makes it a policy to include a copy
of every legal tool that at least some content in this repository is
available under.

Since this repo already contains CC0’d content, this commit adds a copy
of CC0. While progress has been made [1] into closing the
CC0-not-available-under-CC0 loophole [2] [3], it’s not fully closed.
Creative Commons assigned someone to one of the issues, and that
assignee said that CC0 __should__ already be available under CC0 [1].
This, unfortunately, isn’t quite the same as Creative Commons outright
saying that CC0 __is__ available under CC0. Additionally, some text was
added to CC’s Licensing Statement for Content and Software Code [4]
which dedicates “the text of all of the notices and accompanying text on
the license pages” to the public domain using CC0. The key words here
are “on the license pages”. CC0 is not a license [5] so that quote,
unfortunately, doesn’t apply to CC0. Ugh.

With that in mind, I decided to still include attribution text for the
portions of CC0 that are (potentially) CC BY 4.0 licensed. It’s probably
not necessary, but I’d rather err on the side of caution.

[1]: <https://github.com/creativecommons/creativecommons.org/issues/1218#issuecomment-1212525987>
[2]: <https://github.com/creativecommons/creativecommons.org/issues/1212>
[3]: <https://github.com/creativecommons/creativecommons.org/issues/1218>
[4]: <https://creativecommons.org/policies#licensing%20statement>
[5]: <https://creativecommons.org/faq/#how-do-cc-licenses-operate>